requirements.txt
requests>=2.25.1
Pillow>=8.1.0